requirements.txt
# BinanceEventTrader项目依赖 - Python 3.12兼容版本
python-binance==1.0.19
pandas>=2.0.0,<2.2.0
numpy>=1.24.0,<2.0.0
asyncio-throttle==1.0.2
python-dotenv==1.0.0
aiohttp>=3.8.0
websockets>=10.0
loguru==0.7.2
# pandas-ta 暂时移除，使用手动计算避免兼容性问题
# pandas-ta>=0.3.14b0 